refactor(cargo-klyra): remove `cargo-generate` dependency (#1281)
* Remove cargo-generate dependency

* Remove openssl dependency through git2

Replaces network-based (e.g. clonging) features of git2 with
those provided by gix. Thereby rustls can be used instead of
openssl.

* Remove `cargo-generate`-specific test configuration

The purpose of the removed configuration was to
prevent test crashes in CI, which were caused by
the inner workings of `cargo-generate` (see the
deleted `CI DEBUG` comment for more info). Thus,
they can now be deleted.

* Apply review suggestions

* Add a prefix to the temporary directories used.
* Clone template repositories with a depth of 1 commit.
* Check the output of a few of the test cases for the parser.

* Improve design of `generate_project`

The different parts of generating the template
and their use of the arguments passed to
`generate_project` are now divided more clearly.

This decreases the size and overhead of the implementation.

The tests checking the paring are also gone. Instead, the
cloning and the copying of templates is tested. The
integration tests should be sufficient for ensuring that
the paring in this implementation remains correct.

* Apply suggestions