test: add an extra property claim test for pro users (#1448)
While implementing the hard limits I could not figure out why the test
kept failing for a pro user. After some debuging I fount that gateway
did not think the API request was coming from a pro user. I thought the
claim serialization / deserialization might be wrong. So I added a test
to confirm this.

That was not the issue, but this test is helpful none the less. Hence,
this PR.